requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==3.2.18
django-crispy-forms==2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0

pandas==1.3.5